fix: Cyclic dependency of collections to items
