+++ src/modules/collection/sagas.ts
@@ -70,7 +70,7 @@ import {
   SAVE_ITEM_SUCCESS,
   SaveItemSuccessAction
 } from 'modules/item/actions'
-import { isValidText } from 'modules/item/utils'
+import { isValidText, toInitializeItems } from 'modules/item/utils'
 import { locations } from 'routing/locations'
 import { getCollectionId } from 'modules/location/selectors'
 import { builder } from 'lib/api/builder'
@@ -82,7 +82,7 @@ import { getName } from 'modules/profile/selectors'
 import { LoginSuccessAction, LOGIN_SUCCESS } from 'modules/identity/actions'
 import { getCollection, getCollectionItems } from './selectors'
 import { Collection } from './types'
-import { isOwner, getCollectionBaseURI, getCollectionSymbol, toInitializeItems } from './utils'
+import { isOwner, getCollectionBaseURI, getCollectionSymbol } from './utils'
 
 export function* collectionSaga() {
   yield takeEvery(FETCH_COLLECTIONS_REQUEST, handleFetchCollectionsRequest)
