+++ src/modules/collection/utils.ts
@@ -1,14 +1,10 @@
-import { Address } from 'web3x-es/address'
-import { toBN } from 'web3x-es/utils'
 import { env, utils } from 'decentraland-commons'
 import { ChainId, getURNProtocol } from '@dcl/schemas'
 import { ContractName, getContract } from 'decentraland-transactions'
 import { Wallet } from 'decentraland-dapps/dist/modules/wallet/types'
 import { Item } from 'modules/item/types'
-import { getMetadata } from 'modules/item/utils'
 import { isEqual, includes } from 'lib/address'
-import { sortByCreatedAt } from 'lib/sort'
-import { InitializeItem, Collection, Access } from './types'
+import { Collection, Access } from './types'
 import { locations } from 'routing/locations'
 
 export function setOnSale(collection: Collection, wallet: Wallet, isOnSale: boolean): Access[] {
@@ -51,19 +47,6 @@ export function getCollectionBaseURI() {
 export function getCollectionSymbol(collection: Collection) {
   const vowelLessName = collection.name.replace(/a|e|i|o|u|\s/g, '')
   return 'DCL-' + vowelLessName.toUpperCase()
-}
-
-export function toInitializeItems(items: Item[]): InitializeItem[] {
-  return items.sort(sortByCreatedAt).map(toInitializeItem)
-}
-
-export function toInitializeItem(item: Item): InitializeItem {
-  return {
-    metadata: getMetadata(item),
-    rarity: item.rarity!.toLowerCase(),
-    price: toBN(item.price || 0),
-    beneficiary: item.beneficiary ? Address.fromString(item.beneficiary) : Address.ZERO
-  }
 }
 
 export function toCollectionObject(collections: Collection[]) {
